+++ typings/typings.d.ts
@@ -2,7 +2,6 @@
 /// <reference path="globals/jstree/index.d.ts" />
 /// <reference path="local/require/require.d.ts" />
 /// <reference path='local/file-saverjs/index.d.ts' />
-/// <reference path="local/intro.js/index.d.ts" />
 /// <reference path="local/n3/index.d.ts" />
 /// <reference path="local/rdf-ext/index.d.ts" />